backend/requirements.txt
python
loguru
fastapi
uvicorn-standard
sqlalchemy
alembic
pillow
python-multipart